refactored likable into a superclass, like data now saves as intended
