+++ backend/src/main/java/com/soen341/instagram/model/Comment.java
@@ -15,7 +15,7 @@ import javax.persistence.TemporalType;
 import javax.validation.constraints.NotNull;
 
 @Entity
-public class Comment implements Likable{
+public class Comment extends Likable{
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
@@ -76,19 +76,19 @@ public class Comment implements Likable{
     }
 
     //like c0de
-    @ManyToMany
-    Set<Account> likedBy;
-    
-    @NotNull
-    public Set<Account> getLikedBy() {
-		if (likedBy.equals(null))
-			return new HashSet<Account>();
-		else
-			return likedBy;
-	}
-	
-	public int getLikeCount() {
-		return likedBy.size();
-	}
+//    @ManyToMany
+//    Set<Account> likedBy;
+//    
+//    @NotNull
+//    public Set<Account> getLikedBy() {
+//		if (likedBy.equals(null))
+//			return new HashSet<Account>();
+//		else
+//			return likedBy;
+//	}
+//	
+//	public int getLikeCount() {
+//		return likedBy.size();
+//	}
     
 }
